[mocks] Don't overflow stack if method expectation for #respond_to? fails
In the process of generating the error message for a failed method expectation,
ErrorGenerator would call Array#join on an array which contained the receiver.
Internally, Array#join calls #respond_to?(:to_str) on the objects which are being
joined into a string.

If the method expectation was for #respond_to?, the internal call to #respond_to?
could cause the expectation to fail again, causing an infinite loop and resulting
stack overflow.

Avoid this by not using Array#join to generate the error message. Fixes #1020.