[mocks] Avoid exception when class method is stubbed in subclass on MRI 1.8.7
Singleton methods cannot be rebound in MRI 1.8.7, even to a subclass. This means
that when a singleton method is defined on a superclass, and then stubbed out on
a subclass, the code used to cache the original method (so it can be used by
'and_call_original', etc) doesn't work. Actually, it doesn't just 'not work' --
it raises a TypeError which causes the example to fail, *even if* the original
method would never have been used.

Catch this exception, and use an alternate implementation instead which works
even on MRI 1.8.7. Unfortunately, this requires us to expose some implementation
details of MethodDouble.

Also add detailed specs which codify the expected behavior of 'receive' when used
on the same class method in both a superclass and subclass.