[core] Handle syntax error in Ripper
Ripper might fail to parse a Ruby source even if the current runtime
parsed it properly, because some versions of JRuby have different
implementation of runtime parser and Ripper parser.

https://github.com/jruby/jruby/issues/2427

So we handle the syntax error in Ripper and fall back to the simple
single line extraction in that case.