[expectations] Fix #handle_matcher for negative expectations
In quick summary, the negation here has allowed the following thing to be occur
in RSpec for a very, very long time:

```
> expect("foo").to eq "foo"
=> true
Good!

> expect("foo").to eq "bar"
RSpec::Expectations::ExpectationNotMetError:
expected: "bar"
     got: "foo"

(compared using ==)
Good!

> expect("foo").not_to eq "bar"
=> false
WRONG. Should return true.

> expect("foo").not_to eq "foo"
RSpec::Expectations::ExpectationNotMetError:
expected: value != "foo"
     got: "foo"

(compared using ==)
Good!
```

The negation has been around since the extraction of rspec-expectations, so
either it was intended at some point in time to indicate a negative matcher
passing (and this code was never corrected) or something funny must have
happened with the numerous refactorings that have occurred in this area over the
years.

The more prominent trail of commits in this area:

https://github.com/rspec/rspec-expectations/commit/3ba4b82ed09ea4edf6691c1007635b3f81512bca
https://github.com/rspec/rspec-expectations/commit/ef101659c1f8f78171e6311e9ae95107b7b5b36f
https://github.com/rspec/rspec-expectations/commit/358ce674a479118cf486d58f8ee7efe2660c9075

Fun to find and fix!