[core] Remove default val provided to arg `exception_presenter`
When initializing a new instance of `FailedExampleNotification` in [1] we always
use [2] to provide a correctly configured instance of `ExceptionPresenter`. Yet
we pass the `exception_presenter` as a default value when instantianting
`FailedExampleNotification`.

It seems like this default value is only used to help us write a spec for
`FailedExampleNotification`.

In this commit I've removed the default value we provide to
`FailedExampleNotification` (since) it's redundant and simply instantiated a
new instance of the `ExceptionPresenter` in the spec for it to use.

1.) http://bit.ly/1X4NIlU
2.) http://bit.ly/1qn32Qv