[core] Fail fast when asked to `config.include` something that is not a module.
Previously, if you did `config.include :not_a_module, :some_meta`,
and defined an example with `:some_meta` metadata, the ruby type
error would be raised when the example was being prepared to run,
and would lead to other errors due to RSpec’s example setup and
teardown not completing. It was quite confusing.

Instead, we fail fast at the point of the mistake.

Also, we apply this to `config.extend` and `config.prepend`
for consistency.