[core] Force ruby 2.2 version to 2.2.5 on travis
Previously, this was set to `2.2` in the expectation that travis would use the
latest of the 2.2.x branch as a result. However, builds appear to be running
with 2.2.0-p0, which contains erroneous warnings that can force variables and
lets to be needlessly renamed.

While this may be a travis bug, I'm explicitly setting the version in the
configuration to ensure that an up-to-date version is picked up in the
meanwhile.