[core] Include SPEC_OPTS values in bisect reproduction command
Previously the following command:

    SPEC_OPTS="--bisect --seed 125" rspec blah_spec.rb

would return a repro command something like

    rspec blah_spec.rb[1:2,1:4]

However, this won't run in the correct order without the seed. This patch
includes SPEC_OPTS in the bisect reproduction command, minus the `--bisect`
flag.

By default, cukes using aruba to run commands are given an environment with
SPEC_OPTS="-r/path/to/features/support/require_expect_syntax_in_aruba_specs.rb",
which ensures that `should` and `should_not` are not used in the examples.

Now that SPEC_OPTS is included in the bisect reproduction command, however,
this setting shows up in the final output, breaking the existing assertions.

To prevent this I've added a new tag, `@with-clean-spec-opts`, and added this
tag to the logic controlling whether the above file is loaded.