[core] Only apply modules to highest level matching group
For the behaviour when apploying helper modules to existing example
groups to be fully consistent with the behaviour for new groups, we
shouldn't apply the module if any of the example group's parent groups
already have the module applied.

Also refactor how we add hooks to existing groups to use the same code.