[core] Re-initializing the reporter when clearing examples.
This was prompted by #2367 and the cumulative reporting out of
deprecations. Semantically, this could be weird! Should clearing
examples also clear the reporter? If the reporter is configured after
initialization through a means other than standard RSpec.configure, that
configuration would be lost. I don't know why you would do that though?

The original intent of clear_examples (from #1700) seems to imply that
reseting the reporters was desirable, so I think this is OK.

It's weird that reporter belongs to configuration. It mixes state with
configuration. This means that we have to do a "partial" configuration
reset when we want to "reset except configuration", which suggests to me
that we'll probably have more issues here in the future.

Reported (among other things) in #2367