[support] Recognize hash as optional arg when optional keyword is present
Given a method foo(a, b = {}, x: false) it should recognize that calling
foo(1, 'a' => 1, 'b' => 2) is a valid method call.

In case when the method has optional arguments and keyword arguments the
ruby interpreter will treat the last Hash as keyword arguments if it
contains Symbol keys. In the example above calling foo(1, a: 1) will
raise an ArgumentError.