[expectations] Improve Hash formatting in `EnglishPhrasing.list`
We're currently treating all Enumerables as an Array in this method,
resulting in a Hash being formatted as a list of two-element key-value
arrays.

This results in unexpected formatting in the default failure message for
a custom matcher when the `expected` value is a Hash.

This commit updates `EnglishPhrasing.list` to no longer format a Hash as
if it were an Array.

Fixes #1192.